third_party/blink/web_tests/fast/dom/inner-text-with-no-renderer.html

<!doctype>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(() => {
  const div = document.createElement('div');
  div.innerHTML = '&quot;Text&quot;<br>new line';
  assert_equals(div.innerText, '"Text"new line');
 },
'innerText for an element without layout object');
</script>
